Create keepalive.yml
Blockchain-app v1 that is live stores it's created wallets through Vercel in a Upstash database. Now because my portfolio site does not get much traffic, the Upstash database is about to get disabled and archived. I need this database to stay alive so I will write a script to get some activity to the database to keep it up and running.